requirements.txt
Flask>=2.0.0
PyYAML>=6.0
pandas>=1.3.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
